requirements.txt
pyautogui
pynput
pillow
numpy
scikit-learn 